Change bnn binarize to not use H. Add clearCallbacks
BinaryLayer binarize method now doesn't use H to be compatible with pytorch bnn.
As keras implementation (and past saved models) still use it, changes will be made to remain compatible with both.

clearCallbacks method was added to Network and LinearLayer to make it easy to test the model with multiple error p's, without reloading everytime.